lib/intel_blt: Release an offset in the allocator on buffer destroy
Blitter library code requires allocator handle only at the execution
time. That means blt_copy_object structure is allocator agnostic and
offsets are taken for block-copy or fast-copy just before operation.
Operations on bo's like close/create requires to reflect on the
allocator state, otherwise allocator will contain outdated
offsets + sizes leading to fail on allocator operations or the exec.

Helpers blt_create_object() and blt_destroy_object() are unaware
of allocator so if in the meantime object was used in block-copy or
fast-copy it has assigned offset in the allocator. Destroying object
with blt_destroy_object() doesn't release this offset what might
cause failures described above.

There're two possibilities how to solve this - to couple blitter
library with allocator or provide additional helpers which will
release offset in the allocator on destroy path. Latter one is
easier to provide so this is a subject of this change.

Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Karolina Stolarek <karolina.stolarek@intel.com>